feat: `Sidebar` as a component in freya-components
