+++ crates/components/src/lib.rs
@@ -15,6 +15,7 @@ mod loader;
 mod network_image;
 mod progress_bar;
 mod scroll_views;
+mod sidebar;
 mod slider;
 mod switch;
 mod table;
@@ -35,6 +36,7 @@ pub use loader::*;
 pub use network_image::*;
 pub use progress_bar::*;
 pub use scroll_views::*;
+pub use sidebar::*;
 pub use slider::*;
 pub use switch::*;
 pub use table::*;
